chore(.eslintignore): do not check shims-env.d.ts with eslint
