+++ .eslintignore
@@ -1,3 +1,3 @@
 next-env.d.ts
-
+shims-env.d.ts
 ecosystem.config.js
